feat: implement VoicesRadioButtonForm component for improved radio button selection handling and error display
